+++ ci/install.sh
@@ -5,8 +5,8 @@ set -e
 function main
 {
     scripts/install/cmake.sh
-    scripts/install/deadbeef.sh v0.7
     scripts/install/deadbeef.sh v1.8
+    scripts/install/deadbeef.sh v1.9
 }
 
 source "$(dirname $0)/run-in-docker.sh"