+++ js/api_tests/src/deadbeef/test_context.js
@@ -9,7 +9,7 @@ class TestContext extends TestContextBase
         config.playerId = 'deadbeef';
 
         const { BEEFWEB_TEST_DEADBEEF_VERSION } = process.env;
-        const version = BEEFWEB_TEST_DEADBEEF_VERSION || 'v1.8';
+        const version = BEEFWEB_TEST_DEADBEEF_VERSION || 'v1.9';
 
         config.playerDirBase = path.join(config.toolsDir, 'deadbeef', version);